[Decomp][Spell] Cohesion split: extract EffectScriptEffect handler to SpellEffectScript.cpp
EffectScriptEffect is the largest single Spell-effect handler (3,475 lines of
per-spell-ID scripted behaviour). Move its body verbatim into a dedicated
SpellEffectScript.cpp translation unit; the declaration stays in Spell.h and the
SpellEffects[] dispatch table is untouched, so it resolves the handler by address
at link time. SpellEffects.cpp 10,708 -> 7,227. No behaviour change.
